// backend/src/common/enums/role.enum.ts
export enum Role {
  SUPER_ADMIN = 'super_admin',
  OPERATIONAL_ADMIN = 'operational_admin',
  ACCOUNTANT = 'accountant',
  BUSINESS_MANAGER = 'business_manager',
  APEX_BUSINESS_MANAGER = 'apex_business_manager',
  SUPERVISOR = 'supervisor',
  LOAN_MANAGER = 'loan_manager',
  BNPL_MANAGER = 'bnpl_manager',
  INVESTMENT_MANAGER = 'investment_manager',
  INDIVIDUAL = 'individual',
  OPERATIONS = 'operations',
}